requirements.txt
alembic==0.9.6
aniso8601==1.3.0
attrs==19.3.0
bcrypt==3.1.4
boto3==1.11.0
botocore==1.14.0
certifi==2019.11.28
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
docutils==0.15.2
enum-compat==0.0.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-QiniuStorage==0.9.5
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gem==0.1.12
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
Mako==1.0.7
MarkupSafe==1.0
more-itertools==7.2.0
psycopg2-binary==2.8.4
pycparser==2.18
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.5
python-dateutil==2.6.1
python-dotenv==0.10.3
python-editor==1.0.3
python-engineio==2.0.2
python-http-client==3.2.1
python-socketio==1.8.4
pytz==2019.3
qiniu==7.2.6
requests==2.22.0
s3transfer==0.3.0
sendgrid==6.1.0
six==1.11.0
SQLAlchemy==1.2.0
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
